+++ jovo-integrations/jovo-asr-azure/src/index.ts
@@ -1,0 +1,2 @@
+export * from './AzureAsr';
+export * from './Interfaces';