+++ jovo-integrations/jovo-nlu-luis/src/index.ts
@@ -1,0 +1,2 @@
+export * from './LuisNlu';
+export * from './Interfaces';