+++ jovo-platforms/jovo-platform-alexa/src/Alexa.ts
@@ -31,6 +31,8 @@ import { AlexaResponseBuilder } from './core/AlexaResponseBuilder';
 import { GadgetControllerPlugin } from './modules/GadgetControllerPlugin';
 import { ProactiveEventPlugin } from './modules/ProactiveEvent';
 import { AplPlugin } from './modules/AplPlugin';
+import { AskFor } from './modules/AskFor';
+import { AmazonPayPlugin } from './modules/AmazonPay';
 
 export interface Config extends ExtensibleConfig {
   allowedSkillIds: string[];
@@ -114,6 +116,8 @@ export class Alexa extends Extensible implements Platform {
       new Cards(),
       new DialogInterface(),
       new ProactiveEventPlugin(),
+      new AskFor(),
+      new AmazonPayPlugin(),
     );
 
     Jovo.prototype.$alexaSkill = undefined;