+++ jovo-platforms/jovo-platform-alexa/src/index.ts
@@ -6,7 +6,9 @@ export { Alexa } from './Alexa';
 export { AlexaSkill } from './core/AlexaSkill';
 export { AlexaTestSuite } from './core/Interfaces';
 export * from './core/AlexaRequest';
+export * from './modules/AmazonPay';
 export * from './services/AlexaReminder';
+export * from './services/AmazonPayAPI';
 
 export { AlexaResponse } from './core/AlexaResponse';
 
@@ -62,6 +64,7 @@ import { AlexaSpeechBuilder } from './core/AlexaSpeechBuilder';
 import { ProactiveEvent } from './modules/ProactiveEvent';
 
 import { Apl } from './modules/AplPlugin';
+import { AmazonPay } from './modules/AmazonPay';
 
 declare module 'jovo-core/dist/src/Jovo' {
   export interface Jovo {
@@ -458,10 +461,41 @@ declare module './core/AlexaSkill' {
   }
 }
 
+// Amazon Pay
+declare module './core/AlexaSkill' {
+  interface AlexaSkill {
+    $pay?: AmazonPay;
+    pay(): AmazonPay | undefined;
+  }
+}
+
 declare module 'jovo-core/dist/src/Interfaces' {
   interface Output {
     Alexa: {
       Directives?: Directive[];
+      AskForPermission?: AskForPermissionDirective;
     };
+  }
+}
+
+export interface AskForPermissionDirective {
+  type: string;
+  name: 'AskFor';
+  payload: {
+    '@type': 'AskForPermissionsConsentRequest';
+    '@version': string;
+    'permissionScope': string;
+  };
+  token?: string;
+}
+
+// Ask For
+declare module './core/AlexaSkill' {
+  interface AlexaSkill {
+    askForPermission(permissionScope: string, token?: string): this;
+    askForReminders(token?: string): this;
+    getPermissionStatus(): string | undefined;
+    hasPermissionAccepted(): boolean;
+    hasPermissionDenied(): boolean;
   }
 }